+++ R/hidden_aliases.R
@@ -1,0 +1,9 @@
+#' @title Internal page for hidden aliases
+#'
+#' @aliases SimilarRtimeParam-class EicCorrelationParam-class
+#' 
+#' @usage NULL
+#' @format NULL
+#' @keywords internal
+#' @docType methods
+hidden_aliases <- NULL